# be-summit/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Summit API Documentation",
        "description": "API documentation for the BE Summit project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://be-summit.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/mitra": {
            "get": {
                "tags": [
                    "Partner (Admin)"
                ],
                "summary": "List all partner (mitra) profiles (Admin)",
                "operationId": "2f13582e2c120e4a97daa0e0932e44e0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partners fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partners fetched successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MitraResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Partner (Admin)"
                ],
                "summary": "Create a new partner (mitra) profile and user account (Admin)",
                "operationId": "ba64625dd4a83d3dbb82ebb3510ffe0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "nama_pemilik",
                                    "telepon",
                                    "alamat",
                                    "status",
                                    "nik",
                                    "rekening_bank",
                                    "nama_rekening",
                                    "bank"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mitra1@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "nama_pemilik": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "telepon": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Raya Summit No. 10"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Pemilik Basecamp Merbabu Indah"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "aktif",
                                        "enum": [
                                            "aktif",
                                            "suspend"
                                        ]
                                    },
                                    "npwp": {
                                        "type": "string",
                                        "example": "12.345.678.9-012.000"
                                    },
                                    "nik": {
                                        "type": "string",
                                        "example": "3201234567890001"
                                    },
                                    "rekening_bank": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "nama_rekening": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "bank": {
                                        "type": "string",
                                        "example": "Bank BCA"
                                    },
                                    "ewallet": {
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partner profile and user account created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MitraResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MitraResource": {
                "title": "Mitra Resource",
                "description": "Partner profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "nama_pemilik": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "telepon": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Raya Summit No. 10"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Pemilik Basecamp Merbabu Indah",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif"
                    },
                    "npwp": {
                        "type": "string",
                        "example": "12.345.678.9-012.000",
                        "nullable": true
                    },
                    "nik": {
                        "type": "string",
                        "example": "3201234567890001"
                    },
                    "rekening_bank": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "nama_rekening": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "bank": {
                        "type": "string",
                        "example": "Bank BCA"
                    },
                    "ewallet": {
                        "type": "string",
                        "example": "081234567890",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-05T12:00:00+07:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jhon Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jhon@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "pendaki"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T15:34:17+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T15:34:17+07:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}